removed infering SQL
