+++ .vscode/launch.json
@@ -7,7 +7,7 @@
         {
             "name": "Python: Current File",
             "type": "python",
-            "request": "launch",
+            "request": "test",
             "program": "${file}",
             "console": "integratedTerminal",
             "justMyCode": false
