+++ mssql_dataframe/core/write/insert.py
@@ -15,7 +15,6 @@ class insert:
         self,
         connection: pyodbc.connect,
         include_metadata_timestamps: bool = False,
-        autoadjust_sql_objects: bool = False,
     ):
         """Class for inserting data into SQL.
 
@@ -23,19 +22,16 @@ class insert:
         ----------
         connection (pyodbc.Connection) : connection for executing statement
         include_metadata_timestamps (bool, default=False) : include metadata timestamps _time_insert & _time_update for write operations
-        autoadjust_sql_objects (bool, default=False) : if True, create SQL tables or alter SQL columns if needed
         """
         self._connection = connection
         self.include_metadata_timestamps = include_metadata_timestamps
-        self.autoadjust_sql_objects = autoadjust_sql_objects
 
-        # max attempts for creating/modifing SQL tables
-        # value of 3 will: add include_metadata_timestamps columns and/or add other columns and/or increase column size
-        self._adjust_sql_attempts = 3
 
-        # handle failures if autoadjust_sql_objects==True
-        self._modify = modify.modify(connection)
+        # create temporary tables for upsert/merging
         self._create = create.create(connection)
+
+        # add include_metadata_timestamps if needed
+        self._modify = modify.modify(connection)
 
     def insert(
         self,
@@ -49,7 +45,6 @@ class insert:
         ----------
         table_name (str) : name of table to insert data into
         dataframe (pandas.DataFrame): tabular data to insert
-        include_metadata_timestamps (bool, default=None) : override for the class initialized parameter autoadjust_sql_objects to include _time_insert column
 
         Returns
         -------
@@ -110,40 +105,33 @@ class insert:
         schema (pandas.DataFrame) : table column specifications and conversion rules
         dataframe (pandas.DataFrame) : input dataframe with optimal values and types for inserting into SQL
         """
-        for _ in range(0, self._adjust_sql_attempts + 1):
-            try:
-                # dataframe values converted according to SQL data type
-                schema, dataframe = conversion.get_schema(
-                    self._connection,
-                    table_name,
-                    dataframe,
-                    additional_columns,
-                )
-                break
-            except (
-                custom_errors.SQLTableDoesNotExist,
-                custom_errors.SQLColumnDoesNotExist,
-                custom_errors.SQLInsufficientColumnSize,
-            ) as failure:
-                cursor.rollback()
-                # dataframe values may be converted according to SQL data type
-                dataframe = _exceptions.handle(
-                    failure,
-                    table_name,
-                    dataframe,
-                    updating_table,
-                    self.autoadjust_sql_objects,
-                    self._modify,
-                    self._create,
-                )
-                cursor.commit()
-            except Exception as err:
-                cursor.rollback()
-                raise err
-        else:
-            raise RecursionError(
-                f"adjust_sql_attempts={self._adjust_sql_attempts} reached"
+        try:
+            schema, dataframe = conversion.get_schema(
+                self._connection,
+                table_name,
+                dataframe,
+                additional_columns,
             )
+        except custom_errors.SQLColumnDoesNotExist as failure:
+            # add metadata_timestamps
+            cursor.rollback()
+            dataframe = _exceptions.add_metadata_timestamps(
+                failure,
+                table_name,
+                dataframe,
+                self._modify,
+            )
+            cursor.commit()
+            # retry data insert
+            schema, dataframe = conversion.get_schema(
+                self._connection,
+                table_name,
+                dataframe,
+                additional_columns,
+            )
+        except Exception as err:
+            cursor.rollback()
+            raise err
 
         return schema, dataframe
