+++ tests/test_core/test_create.py
@@ -258,311 +258,3 @@ def test_table_sqlpk(sql):
     assert all(schema["odbc_type"] == [pyodbc.SQL_INTEGER, pyodbc.SQL_VARCHAR])
     assert all(schema["odbc_size"] == [4, 0])
     assert all(schema["odbc_precision"] == [0, 0])
-
-
-def test_table_from_dataframe_simple(sql, caplog):
-
-    table_name = "##test_table_from_dataframe_simple"
-    dataframe = pd.DataFrame({"ColumnA": [1]})
-    dataframe = sql.create.table_from_dataframe(table_name, dataframe)
-
-    schema, _ = conversion.get_schema(sql.connection, table_name)
-
-    assert len(schema) == 1
-    assert all(schema.index == "ColumnA")
-    assert all(schema["sql_type"] == "tinyint")
-    assert all(~schema["is_nullable"])
-    assert all(~schema["ss_is_identity"])
-    assert all(schema["pk_seq"].isna())
-    assert all(schema["pk_name"].isna())
-    assert all(schema["pandas_type"] == "UInt8")
-    assert all(schema["odbc_type"] == pyodbc.SQL_TINYINT)
-    assert all(schema["odbc_size"] == 1)
-    assert all(schema["odbc_precision"] == 0)
-
-    result = conversion.read_values(
-        f"SELECT * FROM {table_name}", schema, sql.connection
-    )
-    assert compare_dfs(result, dataframe)
-
-    # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert "Created table" in caplog.record_tuples[0][2]
-
-
-def test_table_from_dataframe_datestr(sql, caplog):
-    table_name = "##test_table_from_dataframe_datestr"
-    dataframe = pd.DataFrame({"ColumnA": ["06/22/2021"]})
-    dataframe = sql.create_meta.table_from_dataframe(table_name, dataframe)
-
-    schema, _ = conversion.get_schema(sql.connection, table_name)
-
-    expected = pd.DataFrame(
-        {
-            "column_name": pd.Series(["ColumnA", "_time_insert"], dtype="string"),
-            "sql_type": pd.Series(["date", "datetime2"], dtype="string"),
-            "is_nullable": pd.Series([False, True]),
-            "ss_is_identity": pd.Series([False, False]),
-            "pk_seq": pd.Series([None, None], dtype="Int64"),
-            "pk_name": pd.Series([None, None], dtype="string"),
-            "pandas_type": pd.Series(
-                ["datetime64[ns]", "datetime64[ns]"], dtype="string"
-            ),
-            "odbc_type": pd.Series(
-                [pyodbc.SQL_TYPE_DATE, pyodbc.SQL_TYPE_TIMESTAMP], dtype="int64"
-            ),
-            "odbc_size": pd.Series([10, 27], dtype="int64"),
-            "odbc_precision": pd.Series([0, 7], dtype="int64"),
-        }
-    ).set_index(keys="column_name")
-    assert compare_dfs(schema[expected.columns], expected)
-
-    result = conversion.read_values(
-        f"SELECT * FROM {table_name}", schema, sql.connection
-    )
-    assert compare_dfs(result[dataframe.columns], dataframe)
-
-    # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert "Created table" in caplog.record_tuples[0][2]
-
-
-def test_table_from_dataframe_errorpk(sql, sample):
-
-    with pytest.raises(ValueError):
-        table_name = "##test_table_from_dataframe_nopk"
-        sql.create.table_from_dataframe(table_name, sample, primary_key="ColumnName")
-
-
-def test_table_from_dataframe_nopk(sql, sample, validation, caplog):
-
-    table_name = "##test_table_from_dataframe_nopk"
-    dataframe = sql.create.table_from_dataframe(
-        table_name, sample.copy(), primary_key=None
-    )
-
-    schema, _ = conversion.get_schema(sql.connection, table_name)
-
-    validation = validation.drop(['_pk','_index'])
-    compare_dfs(schema[validation.columns], validation.loc[schema.index])
-
-    result = conversion.read_values(
-        f"SELECT * FROM {table_name}", schema, sql.connection
-    )
-    assert compare_dfs(result[dataframe.columns], dataframe)
-
-    # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert "Created table" in caplog.record_tuples[0][2]
-
-
-def test_table_from_dataframe_sqlpk(sql, sample, validation, caplog):
-
-    table_name = "##test_table_from_dataframe_sqlpk"
-    dataframe = sql.create.table_from_dataframe(
-        table_name, sample.copy(), primary_key="sql"
-    )
-
-    schema, _ = conversion.get_schema(sql.connection, table_name)
-
-    validation = validation.drop('_index')
-    assert compare_dfs(schema[validation.columns], validation.loc[schema.index])
-    assert pd.notna(schema.at["_pk", "pk_name"])
-    assert schema.loc[schema.index != "_pk", "pk_name"].isna().all()
-
-    result = conversion.read_values(
-        f"SELECT * FROM {table_name}", schema, sql.connection
-    )
-    result = result.reset_index(drop=True)
-    assert compare_dfs(result[dataframe.columns], dataframe)
-
-    # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert "Created table" in caplog.record_tuples[0][2]
-
-
-def test_table_from_dataframe_indexpk_unnamed(sql, sample, validation, caplog):
-
-    table_name = "##test_table_from_dataframe_indexpk_unnamed"
-    dataframe = sql.create.table_from_dataframe(
-        table_name, sample.copy(), primary_key="index"
-    )
-
-    schema, _ = conversion.get_schema(sql.connection, table_name)
-
-    validation = validation.drop('_pk')
-    assert compare_dfs(schema[validation.columns], validation.loc[schema.index])
-    assert pd.notna(schema.at["_index", "pk_name"])
-    assert schema.loc[schema.index != "_index", "pk_name"].isna().all()
-
-    result = conversion.read_values(
-        f"SELECT * FROM {table_name}", schema, sql.connection
-    )
-    assert compare_dfs(result[dataframe.columns], dataframe)
-
-    # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert "Created table" in caplog.record_tuples[0][2]
-
-
-def test_table_from_dataframe_indexpk_named(sql, sample, validation, caplog):
-
-    table_name = "##test_table_from_dataframe_indexpk_named"
-    sample.index.name = "NamedIndex"
-    dataframe = sql.create.table_from_dataframe(
-        table_name, sample.copy(), primary_key="index"
-    )
-
-    schema, _ = conversion.get_schema(sql.connection, table_name)
-
-    validation = validation.drop('_pk')
-    index = validation.index.values
-    index[index=='_index'] = 'NamedIndex'
-    validation.index = index
-    assert compare_dfs(schema[validation.columns], validation.loc[schema.index])
-    assert pd.notna(schema.at["NamedIndex", "pk_name"])
-    assert schema.loc[schema.index != "NamedIndex", "pk_name"].isna().all()
-
-    result = conversion.read_values(
-        f"SELECT * FROM {table_name}", schema, sql.connection
-    )
-    assert compare_dfs(result[dataframe.columns], dataframe)
-
-    # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert "Created table" in caplog.record_tuples[0][2]
-
-
-def test_table_from_dataframe_inferpk_integer(sql, caplog):
-
-    table_name = "##test_table_from_dataframe_inferpk_integer"
-    dataframe = pd.DataFrame(
-        {
-            "_varchar1": ["a", "b", "c", "d", "e"],
-            "_varchar2": ["aa", "b", "c", "d", "e"],
-            "_tinyint": [None, 2, 3, 4, 5],
-            "_smallint": [265, 2, 6, 4, 5],
-            "_int": [32768, 2, 3, 4, 5],
-            "_float1": [1.1111, 2, 3, 4, 5],
-            "_float2": [1.1111, 2, 3, 4, 6],
-        }
-    )
-    dataframe = sql.create.table_from_dataframe(
-        table_name, dataframe, primary_key="infer"
-    )
-
-    schema, _ = conversion.get_schema(sql.connection, table_name)
-
-    assert schema.at["_smallint", "pk_seq"] == 1
-    assert all(schema.loc[schema.index != "_smallint", "pk_seq"].isna())
-
-    result = conversion.read_values(
-        f"SELECT * FROM {table_name}", schema, sql.connection
-    )
-    assert compare_dfs(result[dataframe.columns], dataframe.sort_index())
-
-    # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert "Created table" in caplog.record_tuples[0][2]
-
-
-def test_table_from_dataframe_inferpk_string(sql, caplog):
-
-    table_name = "##test_table_from_dataframe_inferpk_string"
-    dataframe = pd.DataFrame(
-        {
-            "_varchar1": ["a", "b", "c", "d", "e"],
-            "_varchar2": ["aa", "b", "c", "d", "e"],
-        }
-    )
-    dataframe = sql.create.table_from_dataframe(
-        table_name, dataframe, primary_key="infer"
-    )
-
-    schema, _ = conversion.get_schema(sql.connection, table_name)
-
-    assert schema.at["_varchar1", "pk_seq"] == 1
-    assert all(schema.loc[schema.index != "_varchar1", "pk_seq"].isna())
-
-    result = conversion.read_values(
-        f"SELECT * FROM {table_name}", schema, sql.connection
-    )
-    assert compare_dfs(result[dataframe.columns], dataframe)
-
-    # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert "Created table" in caplog.record_tuples[0][2]
-
-
-def test_table_from_dataframe_inferpk_none(sql, caplog):
-
-    table_name = "##test_table_from_dataframe_inferpk_none"
-    dataframe = pd.DataFrame(
-        {
-            "_varchar1": [None, "b", "c", "d", "e"],
-            "_varchar2": [None, "b", "c", "d", "e"],
-        }
-    )
-
-    dataframe = sql.create.table_from_dataframe(
-        table_name, dataframe, primary_key="infer"
-    )
-
-    schema, _ = conversion.get_schema(sql.connection, table_name)
-
-    assert all(schema["pk_seq"].isna())
-
-    result = conversion.read_values(
-        f"SELECT * FROM {table_name}", schema, sql.connection
-    )
-    assert compare_dfs(result[dataframe.columns], dataframe)
-
-    # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert "Created table" in caplog.record_tuples[0][2]
-
-
-def test_table_from_dataframe_composite_pk(sql, caplog):
-
-    table_name = "##test_table_from_dataframe_composite_pk"
-    dataframe = pd.DataFrame(
-        {"ColumnA": [1, 2], "ColumnB": ["a", "b"], "ColumnC": [3, 4]}
-    )
-    dataframe = dataframe.set_index(keys=["ColumnA", "ColumnB"])
-    dataframe = sql.create.table_from_dataframe(
-        table_name, dataframe, primary_key="index"
-    )
-
-    schema, _ = conversion.get_schema(sql.connection, table_name)
-
-    assert schema.at["ColumnA", "pk_seq"] == 1
-    assert schema.at["ColumnB", "pk_seq"] == 2
-
-    result = conversion.read_values(
-        f"SELECT * FROM {table_name}", schema, sql.connection
-    )
-    assert compare_dfs(result[dataframe.columns], dataframe)
-
-    # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert "Created table" in caplog.record_tuples[0][2]
