+++ tests/test_core/test_write/test_merge.py
@@ -6,7 +6,7 @@ import pandas as pd
 
 from mssql_dataframe.connect import connect
 from mssql_dataframe.core import custom_errors, create, conversion
-from mssql_dataframe.core.write import merge
+from mssql_dataframe.core.write import insert, merge
 from mssql_dataframe.__equality__ import compare_dfs
 
 pd.options.mode.chained_assignment = "raise"
@@ -16,6 +16,7 @@ class package:
     def __init__(self, connection):
         self.connection = connection.connection
         self.create = create.create(self.connection)
+        self.insert = insert.insert(self.connection)
         self.merge = merge.merge(self.connection)
         self.merge_meta = merge.merge(self.connection, include_metadata_timestamps=True)
 
@@ -27,44 +28,12 @@ def sql():
     db.connection.close()
 
 
-def test_merge_errors(sql):
-
-    table_name = "##test_merge_errors"
-    sql.create.table(
-        table_name, columns={"ColumnA": "TINYINT", "ColumnB": "VARCHAR(1)"}
-    )
-
-    with pytest.raises(custom_errors.SQLTableDoesNotExist):
-        sql.merge.merge("error" + table_name, dataframe=pd.DataFrame({"ColumnA": [1]}))
-
-    with pytest.raises(custom_errors.SQLColumnDoesNotExist):
-        sql.merge.merge(
-            table_name,
-            dataframe=pd.DataFrame({"ColumnA": [0], "ColumnC": [1]}),
-            match_columns=["ColumnA"],
-        )
-
-    with pytest.raises(custom_errors.SQLInsufficientColumnSize):
-        sql.merge.merge(
-            table_name,
-            dataframe=pd.DataFrame({"ColumnA": [100000], "ColumnB": ["aaa"]}),
-            match_columns=["ColumnA"],
-        )
-
-    with pytest.raises(ValueError):
-        sql.merge.merge(
-            table_name,
-            dataframe=pd.DataFrame({"ColumnA": [100000], "ColumnB": ["aaa"]}),
-            upsert=True,
-            delete_requires=["ColumnB"],
-        )
-
-
 def test_merge_upsert(sql, caplog):
 
     table_name = "##test_merge_upsert"
-    dataframe = pd.DataFrame({"ColumnA": [3, 4]})
-    sql.create.table_from_dataframe(table_name, dataframe, primary_key="index")
+    dataframe = pd.DataFrame({"ColumnA": [3, 4]}, index=pd.Series([0, 1], name='_index'))
+    sql.create.table(table_name, {'ColumnA': 'TINYINT', '_index': 'TINYINT'}, primary_key_column='_index')
+    dataframe = sql.insert.insert(table_name, dataframe)
 
     # delete, but keep in SQL since upserting
     dataframe = dataframe[dataframe.index != 0].copy()
@@ -91,19 +60,15 @@ def test_merge_upsert(sql, caplog):
     assert "_time_insert" not in result.columns
 
     # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert f"Created table: {table_name}" in caplog.record_tuples[0][2]
+    assert len(caplog.record_tuples) == 0
 
 
 def test_merge_one_match_column(sql, caplog):
 
     table_name = "##test_merge_one_match_column"
-    dataframe = pd.DataFrame({"ColumnA": [3, 4]})
-    dataframe = sql.create.table_from_dataframe(
-        table_name, dataframe, primary_key="index"
-    )
+    dataframe = pd.DataFrame({"ColumnA": [3, 4]}, index = pd.Series([0, 1], name='_index'))
+    sql.create.table(table_name, {'ColumnA': 'TINYINT', '_index': 'TINYINT'}, primary_key_column='_index')
+    dataframe = sql.insert.insert(table_name, dataframe)
 
     # delete
     dataframe = dataframe[dataframe.index != 0]
@@ -129,20 +94,17 @@ def test_merge_one_match_column(sql, caplog):
     assert all(result["_time_insert"].notna() == [False, True])
 
     # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 3
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
+    assert len(caplog.record_tuples) == 2
+    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[0][1] == logging.WARNING
-    assert f"Created table: {table_name}" in caplog.record_tuples[0][2]
+    assert (
+        caplog.record_tuples[0][2]
+        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
+    )
     assert caplog.record_tuples[1][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[1][1] == logging.WARNING
     assert (
         caplog.record_tuples[1][2]
-        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
-    )
-    assert caplog.record_tuples[2][0] == "mssql_dataframe.core.write._exceptions"
-    assert caplog.record_tuples[2][1] == logging.WARNING
-    assert (
-        caplog.record_tuples[2][2]
         == f"Creating column '_time_insert' in table '{table_name}' with data type 'datetime2'."
     )
 
@@ -150,10 +112,10 @@ def test_merge_one_match_column(sql, caplog):
 def test_merge_override_timestamps(sql, caplog):
 
     table_name = "##test_merge_override_timestamps"
-    dataframe = pd.DataFrame({"ColumnA": [3, 4]})
-    dataframe = sql.create.table_from_dataframe(
-        table_name, dataframe, primary_key="index"
-    )
+    dataframe = pd.DataFrame({"ColumnA": [3, 4]}, index = pd.Series([0, 1], name='_index'))
+    sql.create.table(table_name, {'ColumnA': 'TINYINT', '_index': 'TINYINT'}, primary_key_column='_index')
+    dataframe = sql.insert.insert(table_name, dataframe)
+
     # update
     dataframe.loc[dataframe.index == 1, "ColumnA"] = 5
 
@@ -169,20 +131,17 @@ def test_merge_override_timestamps(sql, caplog):
     assert all(result["_time_insert"].notna() == [False, False])
 
     # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 3
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
+    assert len(caplog.record_tuples) == 2
+    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[0][1] == logging.WARNING
-    assert f"Created table: {table_name}" in caplog.record_tuples[0][2]
+    assert (
+        caplog.record_tuples[0][2]
+        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
+    )
     assert caplog.record_tuples[1][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[1][1] == logging.WARNING
     assert (
         caplog.record_tuples[1][2]
-        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
-    )
-    assert caplog.record_tuples[2][0] == "mssql_dataframe.core.write._exceptions"
-    assert caplog.record_tuples[2][1] == logging.WARNING
-    assert (
-        caplog.record_tuples[2][2]
         == f"Creating column '_time_insert' in table '{table_name}' with data type 'datetime2'."
     )
 
@@ -191,11 +150,11 @@ def test_merge_two_match_columns(sql, caplog):
 
     table_name = "##test_merge_two_match_columns"
     dataframe = pd.DataFrame(
-        {"State": ["A", "B"], "ColumnA": [3, 4], "ColumnB": ["a", "b"]}
+        {"State": ["A", "B"], "ColumnA": [3, 4], "ColumnB": ["a", "b"]},
+        index = pd.Series([0, 1], name='_index')
     )
-    dataframe = sql.create.table_from_dataframe(
-        table_name, dataframe, primary_key="index"
-    )
+    sql.create.table(table_name, {'State': 'CHAR(1)', 'ColumnA': 'TINYINT', 'ColumnB': 'CHAR(1)', '_index': 'TINYINT'}, primary_key_column='_index')
+    dataframe = sql.insert.insert(table_name, dataframe)
 
     # delete
     dataframe = dataframe[dataframe.index != 0]
@@ -226,20 +185,17 @@ def test_merge_two_match_columns(sql, caplog):
     assert all(result["_time_insert"].notna() == [False, True])
 
     # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 3
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
+    assert len(caplog.record_tuples) == 2
+    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[0][1] == logging.WARNING
-    assert f"Created table: {table_name}" in caplog.record_tuples[0][2]
+    assert (
+        caplog.record_tuples[0][2]
+        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
+    )
     assert caplog.record_tuples[1][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[1][1] == logging.WARNING
     assert (
         caplog.record_tuples[1][2]
-        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
-    )
-    assert caplog.record_tuples[2][0] == "mssql_dataframe.core.write._exceptions"
-    assert caplog.record_tuples[2][1] == logging.WARNING
-    assert (
-        caplog.record_tuples[2][2]
         == f"Creating column '_time_insert' in table '{table_name}' with data type 'datetime2'."
     )
 
@@ -250,7 +206,8 @@ def test_merge_non_pk_column(sql, caplog):
     dataframe = pd.DataFrame(
         {"State": ["A", "B"], "ColumnA": [3, 4], "ColumnB": ["a", "b"]}
     )
-    dataframe = sql.create.table_from_dataframe(table_name, dataframe, primary_key=None)
+    sql.create.table(table_name, {'State': 'CHAR(1)', 'ColumnA': 'TINYINT', 'ColumnB': 'CHAR(1)'})
+    dataframe = sql.insert.insert(table_name, dataframe)
 
     # delete
     dataframe = dataframe[dataframe.index != 0]
@@ -280,20 +237,17 @@ def test_merge_non_pk_column(sql, caplog):
     assert compare_dfs(result[dataframe.columns], dataframe)
 
     # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 3
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
+    assert len(caplog.record_tuples) == 2
+    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[0][1] == logging.WARNING
-    assert f"Created table: {table_name}" in caplog.record_tuples[0][2]
+    assert (
+        caplog.record_tuples[0][2]
+        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
+    )
     assert caplog.record_tuples[1][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[1][1] == logging.WARNING
     assert (
         caplog.record_tuples[1][2]
-        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
-    )
-    assert caplog.record_tuples[2][0] == "mssql_dataframe.core.write._exceptions"
-    assert caplog.record_tuples[2][1] == logging.WARNING
-    assert (
-        caplog.record_tuples[2][2]
         == f"Creating column '_time_insert' in table '{table_name}' with data type 'datetime2'."
     )
 
@@ -304,9 +258,8 @@ def test_merge_composite_pk(sql, caplog):
     dataframe = pd.DataFrame(
         {"State": ["A", "B"], "ColumnA": [3, 4], "ColumnB": ["a", "b"]}
     ).set_index(keys=["State", "ColumnA"])
-    dataframe = sql.create.table_from_dataframe(
-        table_name, dataframe, primary_key="index"
-    )
+    sql.create.table(table_name, {'State': 'CHAR(1)', 'ColumnA': 'TINYINT', 'ColumnB': 'CHAR(1)'}, primary_key_column=["State", "ColumnA"])
+    dataframe = sql.insert.insert(table_name, dataframe)
 
     # delete
     dataframe = dataframe[dataframe.index != ("A", 3)].copy()
@@ -332,10 +285,7 @@ def test_merge_composite_pk(sql, caplog):
     assert "_time_insert" not in result
 
     # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 1
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
-    assert caplog.record_tuples[0][1] == logging.WARNING
-    assert f"Created table: {table_name}" in caplog.record_tuples[0][2]
+    assert len(caplog.record_tuples) == 0
 
 
 def test_merge_one_delete_condition(sql, caplog):
@@ -343,12 +293,10 @@ def test_merge_one_delete_condition(sql, caplog):
     table_name = "##test_merge_one_delete_condition"
     dataframe = pd.DataFrame(
         {"State": ["A", "B", "B"], "ColumnA": [3, 4, 4], "ColumnB": ["a", "b", "b"]},
-        index=[0, 1, 2],
+        index=pd.Series([0, 1, 2], name='_pk'),
     )
-    dataframe.index.name = "_pk"
-    dataframe = sql.create.table_from_dataframe(
-        table_name, dataframe, primary_key="index"
-    )
+    sql.create.table(table_name, {'State': 'CHAR(1)', 'ColumnA': 'TINYINT', 'ColumnB': 'CHAR(1)', '_pk': 'TINYINT'}, primary_key_column="_pk")
+    dataframe = sql.insert.insert(table_name, dataframe)
 
     # delete 2 records
     dataframe = dataframe[dataframe.index == 1].copy()
@@ -385,20 +333,17 @@ def test_merge_one_delete_condition(sql, caplog):
     assert all(result["_time_insert"].notna() == [False, False, True])
 
     # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 3
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
+    assert len(caplog.record_tuples) == 2
+    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[0][1] == logging.WARNING
-    assert f"Created table: {table_name}" in caplog.record_tuples[0][2]
+    assert (
+        caplog.record_tuples[0][2]
+        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
+    )
     assert caplog.record_tuples[1][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[1][1] == logging.WARNING
     assert (
         caplog.record_tuples[1][2]
-        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
-    )
-    assert caplog.record_tuples[2][0] == "mssql_dataframe.core.write._exceptions"
-    assert caplog.record_tuples[2][1] == logging.WARNING
-    assert (
-        caplog.record_tuples[2][2]
         == f"Creating column '_time_insert' in table '{table_name}' with data type 'datetime2'."
     )
 
@@ -413,12 +358,10 @@ def test_merge_two_delete_requires(sql, caplog):
             "ColumnA": [3, 4, 4],
             "ColumnB": ["a", "b", "b"],
         },
-        index=[0, 1, 2],
+        index=pd.Series([0, 1, 2], name='_pk'),
     )
-    dataframe.index.name = "_pk"
-    dataframe = sql.create.table_from_dataframe(
-        table_name, dataframe, primary_key="index"
-    )
+    sql.create.table(table_name, {'State1': 'CHAR(1)', 'State2': 'CHAR(1)', 'ColumnA': 'TINYINT', 'ColumnB': 'CHAR(1)', '_pk': 'TINYINT'}, primary_key_column="_pk")
+    dataframe = sql.insert.insert(table_name, dataframe)
 
     # delete 2 records
     dataframe = dataframe[dataframe.index == 1].copy()
@@ -462,19 +405,16 @@ def test_merge_two_delete_requires(sql, caplog):
     assert all(result["_time_insert"].notna() == [False, False, True])
 
     # assert warnings raised by logging after all other tasks
-    assert len(caplog.record_tuples) == 3
-    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.create"
+    assert len(caplog.record_tuples) == 2
+    assert caplog.record_tuples[0][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[0][1] == logging.WARNING
-    assert f"Created table: {table_name}" in caplog.record_tuples[0][2]
+    assert (
+        caplog.record_tuples[0][2]
+        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
+    )
     assert caplog.record_tuples[1][0] == "mssql_dataframe.core.write._exceptions"
     assert caplog.record_tuples[1][1] == logging.WARNING
     assert (
         caplog.record_tuples[1][2]
-        == f"Creating column '_time_update' in table '{table_name}' with data type 'datetime2'."
-    )
-    assert caplog.record_tuples[2][0] == "mssql_dataframe.core.write._exceptions"
-    assert caplog.record_tuples[2][1] == logging.WARNING
-    assert (
-        caplog.record_tuples[2][2]
         == f"Creating column '_time_insert' in table '{table_name}' with data type 'datetime2'."
     )
